Code Modularization
Modularized and cleaned the code
Added a command module system
