+++ config.ini
@@ -1,5 +1,3 @@
 [General]
 gemini_api_key = your_api_key
 discord_token = your_bot_token
-
-
