+++ config.py
@@ -1,0 +1,8 @@
+import configparser
+
+config_file = 'config.ini'
+config = configparser.ConfigParser(inline_comment_prefixes=('#', ';'))
+config.read(config_file)
+
+GEMINI_API_KEY = config['General']['gemini_api_key']
+DISCORD_TOKEN = config['General']['discord_token']
